Add main components
